+++ opencti-platform/opencti-graphql/src/rules/indicate-sighted/IndicateSightedDefinition.ts
@@ -1,15 +1,9 @@
 import { RELATION_INDICATES } from '../../schema/stixCoreRelationship';
 import type { RuleDefinition } from '../../types/rules';
 import { STIX_SIGHTING_RELATIONSHIP } from '../../schema/stixSightingRelationship';
-import {
-  ENTITY_TYPE_CAMPAIGN,
-  ENTITY_TYPE_INCIDENT,
-  ENTITY_TYPE_INDICATOR,
-  ENTITY_TYPE_INTRUSION_SET,
-  ENTITY_TYPE_MALWARE,
-  ENTITY_TYPE_THREAT_ACTOR_GROUP
-} from '../../schema/stixDomainObject';
+import { ENTITY_TYPE_CAMPAIGN, ENTITY_TYPE_INCIDENT, ENTITY_TYPE_INTRUSION_SET, ENTITY_TYPE_MALWARE, ENTITY_TYPE_THREAT_ACTOR_GROUP } from '../../schema/stixDomainObject';
 import { ENTITY_TYPE_IDENTITY, ENTITY_TYPE_LOCATION } from '../../schema/general';
+import { ENTITY_TYPE_INDICATOR } from '../../modules/indicator/indicator-types';
 
 const id = 'indicate_sighted';
 const name = 'Inference of targeting via a sighting';
