+++ opencti-platform/opencti-graphql/src/rules/observed-sighting/ObserveSightingDefinition.ts
@@ -1,7 +1,8 @@
 import { RELATION_BASED_ON } from '../../schema/stixCoreRelationship';
 import { ABSTRACT_STIX_CYBER_OBSERVABLE } from '../../schema/general';
-import { ENTITY_TYPE_CONTAINER_OBSERVED_DATA, ENTITY_TYPE_INDICATOR } from '../../schema/stixDomainObject';
+import { ENTITY_TYPE_CONTAINER_OBSERVED_DATA } from '../../schema/stixDomainObject';
 import type { RuleDefinition, RuleFilters, RuleScope } from '../../types/rules';
+import { ENTITY_TYPE_INDICATOR } from '../../modules/indicator/indicator-types';
 
 const id = 'observe_sighting';
 const name = 'Sightings of observables via observed data';
