+++ opencti-platform/opencti-graphql/src/rules/report-refs-indicator-based-on/ReportRefIndicatorBasedOnDefinition.ts
@@ -1,7 +1,8 @@
 import { RELATION_BASED_ON } from '../../schema/stixCoreRelationship';
 import { ABSTRACT_STIX_CYBER_OBSERVABLE } from '../../schema/general';
-import { ENTITY_TYPE_CONTAINER_REPORT, ENTITY_TYPE_INDICATOR } from '../../schema/stixDomainObject';
+import { ENTITY_TYPE_CONTAINER_REPORT } from '../../schema/stixDomainObject';
 import type { RuleDefinition, RuleFilters, RuleScope } from '../../types/rules';
+import { ENTITY_TYPE_INDICATOR } from '../../modules/indicator/indicator-types';
 
 const id = 'report_ref_indicator_based_on';
 const name = 'Observables propagation in reports';
