+++ opencti-platform/opencti-graphql/src/rules/report-refs-indicator-based-on/ReportRefIndicatorBasedOnRule.ts
@@ -1,9 +1,10 @@
 /* eslint-disable camelcase */
-import { ENTITY_TYPE_CONTAINER_REPORT, ENTITY_TYPE_INDICATOR } from '../../schema/stixDomainObject';
+import { ENTITY_TYPE_CONTAINER_REPORT } from '../../schema/stixDomainObject';
 import { ABSTRACT_STIX_CYBER_OBSERVABLE } from '../../schema/general';
 import def from './ReportRefIndicatorBasedOnDefinition';
 import buildContainerRefsRule from '../containerWithRefsBuilder';
 import { RELATION_BASED_ON } from '../../schema/stixCoreRelationship';
+import { ENTITY_TYPE_INDICATOR } from '../../modules/indicator/indicator-types';
 
 const ReportRefsIndicatorBasedOnRule = buildContainerRefsRule(def, ENTITY_TYPE_CONTAINER_REPORT, {
   leftType: ENTITY_TYPE_INDICATOR,
