+++ opencti-platform/opencti-graphql/src/rules/report-refs-observable-based-on/ReportRefObservableBasedOnRule.ts
@@ -1,9 +1,10 @@
 /* eslint-disable camelcase */
-import { ENTITY_TYPE_CONTAINER_REPORT, ENTITY_TYPE_INDICATOR } from '../../schema/stixDomainObject';
+import { ENTITY_TYPE_CONTAINER_REPORT } from '../../schema/stixDomainObject';
 import { ABSTRACT_STIX_CYBER_OBSERVABLE } from '../../schema/general';
 import def from './ReportRefObservableBasedOnDefinition';
 import buildContainerRefsRule from '../containerWithRefsBuilder';
 import { RELATION_BASED_ON } from '../../schema/stixCoreRelationship';
+import { ENTITY_TYPE_INDICATOR } from '../../modules/indicator/indicator-types';
 
 const ReportRefsObservableBasedOnRule = buildContainerRefsRule(def, ENTITY_TYPE_CONTAINER_REPORT, {
   leftType: ENTITY_TYPE_INDICATOR,
