+++ opencti-platform/opencti-graphql/src/rules/sighting-observable/SightingObservableRule.ts
@@ -12,11 +12,11 @@ import { createRuleContent } from '../rules';
 import { createInferredRelation, deleteInferredRuleElement } from '../../database/middleware';
 import { listAllRelations, type RelationOptions } from '../../database/middleware-loader';
 import { RELATION_BASED_ON } from '../../schema/stixCoreRelationship';
-import { ENTITY_TYPE_INDICATOR } from '../../schema/stixDomainObject';
 import type { RuleRuntime } from '../../types/rules';
 import { ENTITY_TYPE_IDENTITY, ENTITY_TYPE_LOCATION } from '../../schema/general';
 import { executionContext, RULE_MANAGER_USER } from '../../utils/access';
 import type { AuthContext } from '../../types/user';
+import { ENTITY_TYPE_INDICATOR } from '../../modules/indicator/indicator-types';
 
 /*
 'If **observable A** is `sighted` in **identity/location B** and '
